refactor: extract runtime session registry helpers
